Add files via upload
create connection to openai using api key and endpoint provided. Creation of shortcut to lakehouse to allow retrieval of data in delta tables and select Case Number and Summary as the main column used by openai. Create functions to connect to fabric, load data, generate summaries and save summaries in delta table with append mode.
Create a pipeline to iterate through the script that generate summaries if not generated on first run due to token rate limit.